ai-app-generator/requirements.txt
typer==0.12.3
jinja2==3.1.4
pydantic==2.8.2
httpx==0.27.0
rich==13.7.1
python-dotenv==1.0.1